make domain sep optional argument to `apply`
This allows us to pass in the precomputed domain separator for a given
permutation.

However, there is no meaningful performance difference, because
`apply` is only called directly in the *remainder* handling of the
SIMD operations. Those make up a small portion of the total runtime.

In addition the code is anyhow already bottlenecked by Poseidon and
Keccak permutations.